add to sites-using-linkita list
I'm goind to write blogs in this repo the following years.

By the way, none of the features tags listing at the bottom of the page, toc at the top and giscus is working on my site. I carefully checked the example in your demo but didn't find where the problem is, may you help me check it?